Add a button to addMember from GroupHome
